make discount functionality in card and order

+++ database/migrations/2020_07_22_151106_create_products_table.php
@@ -24,7 +24,7 @@ class CreateProductsTable extends Migration
             $table->string('image')->nullable();
             $table->string('demo')->nullable();
             $table->unsignedBigInteger('price');
-            $table->unsignedBigInteger('off')->nullable();
+            $table->tinyInteger('off')->nullable();
             $table->time('duration')->nullable();
             $table->year('year');
             $table->unsignedBigInteger('views')->default(0);
